Move shared client tests to a dedicated file.
The TestBasics objects have been moved to the utilities, and the shared
tests have been moved to test_client.rs.
